requirements.txt
# Core Django packages
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1

# Database
psycopg2-binary==2.9.10

# Configuration
django-environ==0.12.0

# API Documentation
drf-yasg==1.21.10

# Filtering and utilities
django-filter==25.1

# Rate limiting and throttling
django-ratelimit==4.1.0

# Data generation for testing
Faker==37.5.3

# Testing dependencies
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality tools
black==24.10.0
isort==5.13.2
flake8==7.1.1
bandit==1.8.0
safety==3.3.0

# Required dependencies
asgiref==3.9.1
pytz==2024.2
sqlparse==0.5.3
PyJWT==2.10.1
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
six==1.16.0
uritemplate==4.1.1
PyYAML==6.0.2
inflection==0.5.1